backend/requirements.txt
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cffi==2.0.0
constantly==23.10.4
cryptography==45.0.7
daphne==4.2.1
Django==5.2.6
django-cors-headers==4.8.0
django-environ==0.12.0
djangorestframework==3.16.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyOpenSSL==25.2.0
service-identity==24.2.0
sqlparse==0.5.3
tomli==2.2.1
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.15.0
zope.interface==8.0
